+++ examples/for-tests/src/testContext.js
@@ -1,0 +1,40 @@
+import { getQueryParams } from "./App";
+
+export function getTestContext() {
+    const ret = {
+        disableDefaultUI: getQueryParams("disableDefaultUI") === "true",
+        thirdPartyRedirectURL: localStorage.getItem("thirdPartyRedirectURL"),
+        authRecipe: window.localStorage.getItem("authRecipe") || "emailpassword",
+        usesDynamicLoginMethods: localStorage.getItem("usesDynamicLoginMethods") === "true",
+        clientRecipeListForDynamicLogin: localStorage.getItem("clientRecipeListForDynamicLogin"),
+        mockLoginMethodsForDynamicLogin: localStorage.getItem("mockLoginMethodsForDynamicLogin"),
+    };
+    return ret;
+}
+export function getEnabledRecipes() {
+    const testContext = getTestContext();
+
+    let enabledRecipes = [];
+
+    if (testContext.usesDynamicLoginMethods) {
+        if (testContext.clientRecipeListForDynamicLogin) {
+            enabledRecipes = JSON.parse(testContext.clientRecipeListForDynamicLogin);
+        } else {
+            enabledRecipes = [
+                "emailpassword",
+                "thirdparty",
+                "thirdpartyemailpassword",
+                "passwordless",
+                "thirdpartypasswordless",
+            ];
+        }
+    } else {
+        if (testContext.authRecipe === "both") {
+            enabledRecipes.push("emailpassword", "thirdparty");
+        } else {
+            enabledRecipes.push(testContext.authRecipe);
+        }
+    }
+
+    return enabledRecipes;
+}