+++ lib/ts/recipe/thirdparty/index.ts
@@ -126,7 +126,6 @@ export default class Wrapper {
     static ActiveDirectory = ActiveDirectory;
     static BoxySAML = BoxySAML;
     static Okta = Okta;
-    static Twitter = Twitter;
     static ComponentsOverrideProvider = RecipeComponentsOverrideContextProvider;
 }