+++ lib/ts/recipe/thirdpartypasswordless/recipe.tsx
@@ -20,6 +20,7 @@
 import ThirdpartyPasswordlessWebJS from "supertokens-web-js/recipe/thirdpartypasswordless";
 
 import { SSR_ERROR } from "../../constants";
+import SuperTokens from "../../superTokens";
 import { isTest } from "../../utils";
 import AuthRecipe from "../authRecipe";
 import Passwordless from "../passwordless/recipe";
@@ -70,12 +71,18 @@ export default class ThirdPartyPasswordless extends AuthRecipe<
             typeof ThirdpartyPasswordlessWebJS
         > = ThirdpartyPasswordlessWebJS
     ) {
+        if (
+            SuperTokens.usesDynamicLoginMethods === false &&
+            config.disablePasswordless === true &&
+            config.thirdpartyConfig?.signInAndUpFeature.providers.length === 0
+        ) {
+            throw new Error("ThirdParty signInAndUpFeature providers array cannot be empty.");
+        }
         super(config);
-
         this.passwordlessRecipe =
             recipes.passwordlessInstance !== undefined
                 ? recipes.passwordlessInstance
-                : this.config.passwordlessConfig === undefined
+                : SuperTokens.usesDynamicLoginMethods === false && this.config.disablePasswordless
                 ? undefined
                 : new Passwordless(
                       {