+++ lib/ts/recipe/thirdpartypasswordless/types.ts
@@ -125,18 +125,17 @@ export type UserInput = (
     linkClickedScreenFeature?: PasswordlessFeatureBaseConfig;
     oAuthCallbackScreen?: FeatureBaseConfig;
     disablePasswordless?: boolean;
-    disableThirdParty?: boolean;
 } & AuthRecipeModuleUserInput<GetRedirectionURLContext, PreAndPostAPIHookAction, OnHandleEventContext>;
 
 export type Config = UserInput &
     AuthRecipeModuleConfig<GetRedirectionURLContext, PreAndPostAPIHookAction, OnHandleEventContext>;
 
 export type NormalisedConfig = {
-    passwordlessConfig: NormalisedPasswordlessConfig | undefined;
+    passwordlessConfig: NormalisedPasswordlessConfig;
     thirdpartyConfig: NormalisedThirdPartyConfig | undefined;
 
     thirdPartyProviderAndEmailOrPhoneFormStyle: string | undefined;
-
+    disablePasswordless: boolean;
     override: {
         functions: (
             originalImplementation: RecipeInterface,